move processAnnotations to sdk
