+++ packages/app/components/permissions/types.ts
@@ -1,25 +1,5 @@
-import { OpenAPIV3 } from "openapi-types"
 import { PermissionCoerced } from "zodiac-roles-sdk"
-
-export type OpenAPIParameter = Omit<OpenAPIV3.ParameterObject, "schema"> & {
-  schema: OpenAPIV3.SchemaObject
-}
-type T = OpenAPIParameter["schema"]
-
-export interface Preset {
-  permissions: PermissionCoerced[]
-  uri: string
-  serverUrl: string
-  apiInfo: OpenAPIV3.InfoObject
-  path: string
-  paramValues: Record<string, string | number | string[] | number[]>
-  operation: {
-    summary?: string
-    description?: string
-    tags?: string[]
-    parameters: OpenAPIParameter[]
-  }
-}
+import { Preset } from "zodiac-roles-sdk/annotations"
 
 export enum DiffFlag {
   Added = "Added",
